Added initial back bone for first time download mocks. Fixed bug when sending sms to multiple contacts.
