+++ app/src/main/java/com/example/dariomolina/alerta/About.java
@@ -1,0 +1,26 @@
+package com.example.dariomolina.alerta;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+
+public class About extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.about_layout);
+
+        ((findViewById(R.id.forwardButton))).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(getApplicationContext(), AboutPermission.class);
+                startActivity(intent);
+                return;
+            };
+        });
+
+    }
+}
