+++ app/src/main/java/com/example/dariomolina/alerta/SMS.java
@@ -21,19 +21,17 @@ import java.util.Arrays;
 public class SMS
 {
 
-    private Context applicationContext;
+    private Activity applicationContext;
     private BroadcastReceiver messageSendReceiver, messageDeliveredReceiver;
     private final String SEND = "SMS_SEND";
     private final String DELIVERED = "SMS_DELIVERED";
     private static final int SMS_MAX_LENGTH = 160;
-    private String name;
 
-    public SMS(Context context)
+    public SMS(Activity activity)
     {
-        applicationContext = context;
+        applicationContext = activity;
         messageSendReceiver = null;
         messageDeliveredReceiver = null;
-        name = null;
     }
 
     public void registerReceivers()
@@ -45,19 +43,19 @@ public class SMS
                 {
                     case Activity.RESULT_OK:
                         Log.d("registerRecievers", "SMS has been sent.");
-                        Toast.makeText(applicationContext, "Mensaje de texto a sido enviado para " + name + ".", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(applicationContext, "Mensaje de texto a sido enviado para " + intent.getExtras().getString("name") + ".", Toast.LENGTH_SHORT).show();
                         break;
                     case SmsManager.RESULT_ERROR_GENERIC_FAILURE:
                         Log.d("registerRecievers", "Generic Test failure. SMS not send.");
-                        Toast.makeText(applicationContext, "Mensaje de texto no fue enviado para " + name + ". Porfavor intente de nuevo.", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(applicationContext, "Mensaje de texto no fue enviado para " + intent.getExtras().getString("name") + ". Porfavor intente de nuevo.", Toast.LENGTH_SHORT).show();
                         break;
                     case SmsManager.RESULT_ERROR_NO_SERVICE:
                         Log.d("registerRecievers", "No Service.  SMS not send.");
-                        Toast.makeText(applicationContext, "Sin Servicio, mensaje de Texto no fue enviado para " + name + ".", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(applicationContext, "Sin Servicio, mensaje de Texto no fue enviado para " + intent.getExtras().getString("name") + ".", Toast.LENGTH_SHORT).show();
                         break;
                     case SmsManager.RESULT_ERROR_NULL_PDU:
                         Log.d("registerRecievers", "Null PDU. SMS not send.");
-                        Toast.makeText(applicationContext, "Mensaje de texto no fue enviado para " + name + ".  Porfavor intente de nuevo.", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(applicationContext, "Mensaje de texto no fue enviado para " + intent.getExtras().getString("name") + ".  Porfavor intente de nuevo.", Toast.LENGTH_SHORT).show();
                         break;
                     case SmsManager.RESULT_ERROR_RADIO_OFF:
                         Log.d("registerRecievers", "Error Radio off");
@@ -75,11 +73,11 @@ public class SMS
                 {
                     case Activity.RESULT_OK:
                         Log.d("registerRecievers", "SMS delivered");
-                        Toast.makeText(applicationContext, "Su contacto " + name + " a recibido su mensaje de texto.", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(applicationContext, "Su contacto " + intent.getExtras().getString("name") + " a recibido su mensaje de texto.", Toast.LENGTH_SHORT).show();
                         break;
                     case Activity.RESULT_CANCELED:
                         Log.d("registerRecievers", "SMS not delivered");
-                        Toast.makeText(applicationContext, "Su contacto " + name + " no recibio su mensaje de texto.", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(applicationContext, "Su contacto " + intent.getExtras().getString("name") + " no recibio su mensaje de texto.", Toast.LENGTH_SHORT).show();
                         break;
                 }
             }
@@ -89,12 +87,12 @@ public class SMS
         applicationContext.registerReceiver(messageDeliveredReceiver , new IntentFilter(DELIVERED));
     }
 
-    public void sendSMS(String phoneNumber, String message, String name)
+    public void sendSMS(String phoneNumber, String message, String name, int id)
     {
-        this.name = name;
         Log.d("sendSMS", "registering pending intents for sms send and  sms delivered.");
-        PendingIntent sentPI = PendingIntent.getBroadcast(applicationContext, 0, new Intent(SEND), 0);
-        PendingIntent deliveredPI = PendingIntent.getBroadcast(applicationContext, 0, new Intent(DELIVERED), 0);
+
+        PendingIntent sentPI = PendingIntent.getBroadcast(applicationContext, id, new Intent(SEND).putExtra("name",name), PendingIntent.FLAG_ONE_SHOT);
+        PendingIntent deliveredPI = PendingIntent.getBroadcast(applicationContext, id, new Intent(DELIVERED).putExtra("name",name), PendingIntent.FLAG_ONE_SHOT);
         SmsManager smsManager = SmsManager.getDefault();
 
         if(message.length() > SMS_MAX_LENGTH) {
